requirements.txt
python-telegram-bot==22.0
python-dotenv==1.0.1
google-generativeai==0.7.0
langdetect==1.0.9
duckduckgo-search==8.0.0
